+++ .gitignore
@@ -131,4 +131,3 @@ dist
 
 .DS_Store
 
-package-lock.json
