+++ client/.gitignore
@@ -19,4 +19,3 @@ Thumbs.db
 ehthumbs.db
 /coverage/
 *.localhistory
-package-lock.json
