+++ package.json
@@ -9,26 +9,27 @@
     "type-check": "tsc --noEmit"
   },
   "dependencies": {
+    "@apollo/client": "^3.7.0",
+    "apollo-server-express": "^3.13.0",
+    "bcryptjs": "^2.4.3",
+    "graphql": "^16.6.0",
+    "jwt-decode": "^3.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.8.0",
-    "graphql": "^16.6.0",
-    "@apollo/client": "^3.7.0",
-    "jwt-decode": "^3.1.2",
-    "bcryptjs": "^2.4.3"
+    "react-router-dom": "^6.8.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "css-loader": "^6.8.1",
+    "eslint": "^8.44.0",
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.3",
     "typescript": "^5.1.3",
     "webpack": "^5.85.0",
     "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.0",
-    "ts-loader": "^9.4.3",
-    "eslint": "^8.44.0",
-    "@typescript-eslint/parser": "^6.10.0",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "css-loader": "^6.8.1",
-    "style-loader": "^3.3.3"
+    "webpack-dev-server": "^4.15.0"
   }
 }
