--- server/src/models/contacts.ts
@@ -1,17 +1,0 @@
-import mongoose, { Schema, Document } from 'mongoose';
-
-export interface IContact extends Document {
-  name: string;
-  email: string;
-  message: string;
-  submittedAt: Date;
-}
-
-const contactSchema = new Schema<IContact>({
-  name: { type: String, required: true },
-  email: { type: String, required: true },
-  message: { type: String, required: true },
-  submittedAt: { type: Date, default: Date.now },
-});
-
-export default mongoose.model<IContact>('Contact', contactSchema);
