--- server/src/models/index.ts
@@ -1,5 +1,0 @@
-import Post from './post.js';
-import User from './user.js';
-
-
-export { User, Post };
